subida 2

+++ src/main/java/org/example/model/entity/Producto.java
@@ -4,7 +4,7 @@ public class Producto {
     private int id;
     private String nombre;
     private String descripcion;
-    private int precio;
+    private double precio;
     private int stock;
     private String imagen;
     private Categoria id_categoria;
@@ -12,7 +12,7 @@ public class Producto {
     public Producto() {
     }
 
-    public Producto(int id, String nombre, String descripcion, int precio, int stock, String imagen, Categoria id_categoria) {
+    public Producto(int id, String nombre, String descripcion, double precio, int stock, String imagen, Categoria id_categoria) {
         this.id = id;
         this.nombre = nombre;
         this.descripcion = descripcion;
@@ -46,11 +46,11 @@ public class Producto {
         this.descripcion = descripcion;
     }
 
-    public int getPrecio() {
+    public double getPrecio() {
         return precio;
     }
 
-    public void setPrecio(int precio) {
+    public void setPrecio(double precio) {
         this.precio = precio;
     }
